Unnecessary tracking of lock file
The `composer.lock` file locks versions of software together, but it is
only relevant to aplications. For libraries, the version declaration
needs to be specific so applications may not consume it in a way it was
not supposed to work on.

Does some house cleaning by providing a proper continuous integration
suite to the whoe thing. The `phpunit.xml.dist` move was made so running
phpunit on the root folder (where most people like to work from) would
use *by default* a configuration file, also ignores the user
`phpunit.xml` so anyone may have rights to their own opinion.

Ignoring the `*.swp` (VIM swap files) is just something I am doing to
teach a fellow user in arms how good can it be to *git globally ignore*
swap files and such. I even had to hear, silently - which typing this
very message on a foreign keyboard, to go on and ignore `*.sw*` all the
way. Hopefully, now, he understands something more.... I hope that it is
how valuable commits can be. ;)

XOXO

PS: You are welcome for the `git add -p`.