Finally something to explain this
This is a first early version of this readme. I think it is because just
now I try to understand wtf it is doing.

Because I project I work I decide to implement some ngram (again) and do
this.
In fact, I just realize it exist when I search form ngram on packagist
for another project.
I hope it is usefull for you.